% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{data}
\format{BED files}
\source{
Genomic Quebec
}
\description{
BS-Seq data for three cell type
}
\details{
The data is part of whole blood from Sweeden.
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
