% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMCHMM-package.R
\docType{package}
\name{DMCHMM-package}
\alias{DMCHMM-package}
\alias{DMCHMM}
\title{Differentially Methylated CpG using Hidden Markov Model}
\description{
DMCHMM is a novel profiling tool for identifying
differentially methylated CpG sites using Hidden Markov Model in bisulfite
sequencing data.
}
\section{\code{DMCHMM} methods}{

\code{\link{BSData}},
\code{\link{BSDMCs}},
\code{\link{methHMEM}},
\code{\link{methHMMCMC}},
\code{\link{findDMCs}},
\code{\link{qqDMCs}},
\code{\link{manhattanDMCs}},
\code{\link{readBismark}},
\code{\link{writeBED}}.
}

\section{\code{DMCHMM} objects}{

\code{\link{BSData-class}},
\code{\link{BSDMCs-class}}
}

