% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methLevels-method.R
\docType{methods}
\name{methLevels-method}
\alias{methLevels-method}
\alias{methLevels-method}
\alias{methLevels,BSDMCs-method}
\alias{methLevels-method}
\alias{methLevels}
\alias{methLevels<-,BSDMCs,matrix-method}
\alias{methLevels-method}
\alias{methLevels<-}
\title{methLevels method}
\usage{
methLevels(object)

methLevels(object) <- value

\S4method{methLevels}{BSDMCs}(object)

\S4method{methLevels}{BSDMCs,matrix}(object) <- value
}
\arguments{
\item{object}{A \code{\link{BSData-class}} or \code{\link{BSDMCs-class}}
object}

\item{value}{An integer matrix}
}
\value{
A matrix

A \code{\link{BSDMCs-class}} object
}
\description{
Returns \code{methLevels} stored in \code{\link{BSDMCs-class}}

Assigns \code{methLevels} to \code{\link{BSDMCs-class}}
}
\examples{
set.seed(1980)
nr <- 200; nc <- 8
metht <- matrix(as.integer(runif(nr * nc, 0, 100)), nr)
methc <- matrix(rbinom(n=nr*nc,c(metht),prob = runif(nr*nc)),nr,nc)
meths <- matrix(as.integer(runif(nr * nc, 0, 10)), nr)
methl <- methc/metht
r1 <- GRanges(rep("chr1", nr), IRanges(1:nr, width=1), strand="*")
names(r1) <- 1:nr
cd1 <- DataFrame(Group=rep(c("G1","G2"),each=nc/2),row.names=LETTERS[1:nc])
OBJ2 <- cBSDMCs(rowRanges=r1,methReads=methc,totalReads=metht,
methLevels=methl,methStates=meths,colData=cd1)
methLevels(OBJ2)
methLevels(OBJ2) <- methl
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
