% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methReads-method.R
\docType{methods}
\name{methReads-method}
\alias{methReads-method}
\alias{methReads-method}
\alias{methReads-method}
\alias{methReads-method}
\alias{methReads,BSData-method}
\alias{methReads-method}
\alias{methReads}
\alias{methReads<-,BSData,matrix-method}
\alias{methReads-method}
\alias{methReads<-}
\alias{methReads,BSDMCs-method}
\alias{methReads-method}
\alias{methReads}
\alias{methReads<-,BSDMCs,matrix-method}
\alias{methReads-method}
\alias{methReads<-}
\title{methReads method}
\usage{
methReads(object)

methReads(object) <- value

methReads(object)

methReads(object) <- value

\S4method{methReads}{BSData}(object)

\S4method{methReads}{BSData,matrix}(object) <- value

\S4method{methReads}{BSDMCs}(object)

\S4method{methReads}{BSDMCs,matrix}(object) <- value
}
\arguments{
\item{object}{A \code{\link{BSData-class}} or \code{\link{BSDMCs-class}}
object}

\item{value}{An integer matrix}
}
\value{
A matrix

A \code{\link{BSData-class}} object

A matrix

A \code{\link{BSDMCs-class}} object
}
\description{
Returns \code{methReads} stored in \code{\link{BSData-class}}

Assigns \code{methReads} to \code{\link{BSData-class}}

Returns \code{methReads} stored in \code{\link{BSDMCs-class}}

Assigns \code{methReads} to \code{\link{BSDMCs-class}}
}
\examples{
nr <- 200; nc <- 8
metht <- matrix(as.integer(runif(nr * nc, 0, 100)), nr)
methc <- matrix(rbinom(n=nr*nc,c(metht),prob = runif(nr*nc)),nr,nc)
r1 <- GRanges(rep("chr1", nr), IRanges(1:nr, width=1), strand="*")
names(r1) <- 1:nr
cd1 <- DataFrame(Group=rep(c("G1","G2"),each=nc/2),row.names=LETTERS[1:nc])
OBJ1 <- BSData(rowRanges=r1,methReads=methc,totalReads=metht,colData=cd1)
methReads(OBJ1)
methReads(OBJ1) <- methc
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
