% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methStates-method.R
\name{methStates-method}
\alias{methStates-method}
\alias{methStates,BSDMCs-method}
\alias{methStates}
\alias{methStates<-,BSDMCs,matrix-method}
\alias{methStates<-}
\title{methStates method}
\usage{
methStates(object)

methStates(object) <- value

\S4method{methStates}{BSDMCs}(object)

\S4method{methStates}{BSDMCs,matrix}(object) <- value
}
\arguments{
\item{object}{A \code{\link{BSData-class}} or \code{\link{BSDMCs-class}}
object}

\item{value}{An integer matrix}
}
\value{
A matrix

A \code{\link{BSDMCs-class}} object
}
\description{
Returns \code{methStates} stored in \code{\link{BSDMCs-class}}

Assigns \code{methStates} to \code{\link{BSDMCs-class}}
}
\examples{
set.seed(1980)
nr <- 150; nc <- 8
metht <- matrix(as.integer(runif(nr * nc, 0, 100)), nr)
methc <- matrix(rbinom(n=nr*nc,c(metht),prob = runif(nr*nc)),nr,nc)
meths <- matrix(as.integer(runif(nr * nc, 0, 10)), nr)
methl <- methc/metht
methv <- matrix((runif(nr * nc, 0.1, 0.5)), nr)
r1 <- GRanges(rep('chr1', nr), IRanges(1:nr, width=1), strand='*')
names(r1) <- 1:nr
cd1 <- DataFrame(Group=rep(c('G1','G2'),each=nc/2),row.names=LETTERS[1:nc])
OBJ2 <- cBSDMCs(rowRanges=r1,methReads=methc,totalReads=metht,
methLevels=methl,methStates=meths,methVars=methv,colData=cd1)
methStates(OBJ2)
methStates(OBJ2)<- meths
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
