% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/readBismark-method.R
\docType{methods}
\name{readBismark-method}
\alias{readBismark-method}
\alias{readBismark,character,DataFrame-method}
\alias{readBismark}
\alias{readBismark,character,data.frame-method}
\alias{readBismark,character,character-method}
\title{readBismark method}
\usage{
readBismark(files, colData)

\S4method{readBismark}{character,DataFrame}(files, colData)

\S4method{readBismark}{character,data.frame}(files, colData)

\S4method{readBismark}{character,character}(files, colData)
}
\arguments{
\item{files}{A character list}

\item{colData}{Object of class \code{"DataFrame"} containing information on
variable values of the samples}
}
\value{
A \code{\link{BSData-class}} object
}
\description{
reads BS-Seq data
}
\examples{
fn <- list.files(system.file('extdata',package = 'DMCHMM'))
fn.f <- list.files(system.file('extdata',package='DMCHMM'), full.names=TRUE)
OBJ <- readBismark(fn.f, fn)
cdOBJ <- DataFrame(Cell = factor(c('BC', 'TC','Mono'),
labels = c('BC', 'TC', 'Mono')), row.names = c('BCU1568','BCU173','BCU551'))
colData(OBJ) <- cdOBJ
OBJ
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
