% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/totalReads-method.R
\docType{methods}
\name{totalReads-method}
\alias{totalReads-method}
\alias{totalReads,BSData-method}
\alias{totalReads}
\alias{totalReads<-,BSData,matrix-method}
\alias{totalReads<-}
\alias{totalReads,BSDMCs-method}
\alias{totalReads<-,BSDMCs,matrix-method}
\title{totalReads method}
\usage{
totalReads(object)

totalReads(object) <- value

totalReads(object)

totalReads(object) <- value

\S4method{totalReads}{BSData}(object)

\S4method{totalReads}{BSData,matrix}(object) <- value

\S4method{totalReads}{BSDMCs}(object)

\S4method{totalReads}{BSDMCs,matrix}(object) <- value
}
\arguments{
\item{object}{A \code{\link{BSData-class}} or \code{\link{BSDMCs-class}}
object}

\item{value}{An integer matrix}
}
\value{
A matrix

A \code{\link{BSData-class}} object

A matrix

A \code{\link{BSDMCs-class}} object
}
\description{
Returns \code{totalReads} stored in \code{\link{BSData-class}}

Assigns \code{totalReads} to \code{\link{BSData-class}}

Returns \code{totalReads} stored in \code{\link{BSDMCs-class}}

Assigns \code{totalReads} to \code{\link{BSDMCs-class}}
}
\examples{
nr <- 150; nc <- 8
metht <- matrix(as.integer(runif(nr * nc, 0, 100)), nr)
methc <- matrix(rbinom(n=nr*nc,c(metht),prob = runif(nr*nc)),nr,nc)
r1 <- GRanges(rep('chr1', nr), IRanges(1:nr, width=1), strand='*')
names(r1) <- 1:nr
cd1 <- DataFrame(Group=rep(c('G1','G2'),each=nc/2),row.names=LETTERS[1:nc])
OBJ1 <- cBSData(rowRanges=r1,methReads=methc,totalReads=metht,colData=cd1)
totalReads(OBJ1)
totalReads(OBJ1) <- metht
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
