% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/writeBED-method.R
\docType{methods}
\name{writeBED-method}
\alias{writeBED-method}
\alias{writeBED,BSData,character,character-method}
\alias{writeBED-method}
\alias{writeBED}
\alias{writeBED,BSData,character,missing-method}
\alias{writeBED-method}
\alias{writeBED}
\alias{writeBED,BSData,missing,character-method}
\alias{writeBED-method}
\alias{writeBED}
\alias{writeBED,BSData,missing,missing-method}
\alias{writeBED-method}
\alias{writeBED}
\alias{writeBED,BSDMCs,character,character-method}
\alias{writeBED-method}
\alias{writeBED}
\alias{writeBED,BSDMCs,character,missing-method}
\alias{writeBED-method}
\alias{writeBED}
\alias{writeBED,BSDMCs,missing,character-method}
\alias{writeBED-method}
\alias{writeBED}
\alias{writeBED,BSDMCs,missing,missing-method}
\alias{writeBED-method}
\alias{writeBED}
\title{writeBED method}
\usage{
writeBED(object, name, file)

\S4method{writeBED}{BSData,character,character}(object, name, file)

\S4method{writeBED}{BSData,character,missing}(object, name)

\S4method{writeBED}{BSData,missing,character}(object, file)

\S4method{writeBED}{BSData,missing,missing}(object)

\S4method{writeBED}{BSDMCs,character,character}(object, name, file)

\S4method{writeBED}{BSDMCs,character,missing}(object, name)

\S4method{writeBED}{BSDMCs,missing,character}(object, file)

\S4method{writeBED}{BSDMCs,missing,missing}(object)
}
\arguments{
\item{object}{A \code{\link{BSData-class}} or \code{\link{BSDMCs-class}}
object}

\item{name}{A character list}

\item{file}{A character}
}
\value{
BED files
}
\description{
write BS-Seq data to BED files
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
