% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{data}
\format{
BED files
}
\source{
Genomic Quebec
}
\description{
A part of BS-Seq data for three cell type:
WGBS data were derived from whole blood collected on a cohort of healthy
individuals from Sweden. Cell lines were separated into T-cells (19 samples),
 monocytes (13 samples) and B-cells (8 samples). Sequencing was performed on
 the Illumina HiSeq2000/2500 system for each of the 40 samples, separately.
 For illustration only 3 samples each containg 30,440 CpG sites around
 BLK gene are provided here. The whole data are analyzed in the cited paper.
}
\details{
The data is part of whole blood from Sweeden.
}
\author{
Farhad Shokoohi <shokoohi@icloud.com>
}
